rework test structure and test that error alert shows when error on patchTask
